refactor: move remaining miner shared helpers to helpers.py (#435)
